% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT_tran.R
\name{GT_tran}
\alias{GT_tran}
\title{Transfer learning in group testing scenario}
\usage{
GT_tran(
  m,
  k,
  pi_0,
  pi_1,
  m_1,
  p_tilde_j,
  v_tilde_j,
  w = 1,
  b1 = 0.01,
  b2 = 0.01,
  eta = NULL,
  lambda = NULL
)
}
\arguments{
\item{m}{Numeric, the group number}

\item{k}{Numeric, the group size}

\item{pi_0}{Numeric, the specificity in group level}

\item{pi_1}{Numeric, the sensitivity in group level}

\item{m_1}{Integer, Number of groups which has been test positive}

\item{p_tilde_j}{Vector, the estimation of p from J source datasets}

\item{v_tilde_j}{Vector, the asymptotic variance of p_tilde_j}

\item{w}{The weight which measure the importance of type 1 error and type 2 error in the first strategy of screening method. Default is 1.}

\item{b1}{Numeric, a small number to evaluate the type 1 error (Default is 0.01)}

\item{b2}{Numeric, a small number to evaluate the type 2 error (Default is 0.01)}

\item{eta}{Numeric, the tolerent parameter of Condition 1 (Default is `\eqn{0.5 + log_m J - log_m \max\limits_{j}\{ \tilde{v}_j/\tilde{v} \} }`)}

\item{lambda}{Numeric, the parameter to evaluate the importance between target information and source information (Default is `\eqn{ \sqrt{\frac{s \tilde{v}}{\max\limits_{j}\{\tilde{v}_j\}}} }`)}
}
\value{
A list contains all the statistics and the estimation after transfer learning.
\item{m}{The group number}
\item{k}{The group size}
\item{pi_0}{The specificity in group level}
\item{pi_1}{The sensitivity in group level}
\item{m_1}{Number of groups which has been test positive}
\item{g_tilde}{The estimation of probability that a group reported to be positive without transfer leaning}
\item{p_tilde}{The estimation of disease prevalence without transfer leaning}
\item{v_tilde}{The variance of p_tilde}
\item{lambda}{The parameter to evaluate the importance between target information and source information}
\item{g_hat}{The estimation of probability that a group reported to be positive after transfer leaning}
\item{p_hat}{The estimation of disease prevalence after transfer leaning}
\item{v_hat}{The variance of p_hat}
}
\description{
GT_tran is used to estimate the disease prevalence in group testing with auxiliary summary statistics.
}
\examples{
tran = GT_tran(500,5,0.95,0.95,41,
c(0.009627190,0.009347203,0.008905579,0.009290185,0.010037513),
c(9.083533e-07,2.552419e-07,6.768645e-07,5.780736e-07,8.428058e-07))
tran
}
