% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GT_tran.R
\name{print.GT_tran}
\alias{print.GT_tran}
\title{Transfer learning in group testing scenario}
\usage{
\method{print}{GT_tran}(x, ...)
}
\arguments{
\item{x}{An object of class `GT_tran`.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print a summary of GT_tran.
}
\examples{
tran = GT_tran(500,5,0.95,0.95,41,
c(0.009627190,0.009347203,0.008905579,0.009290185,0.010037513),
c(9.083533e-07,2.552419e-07,6.768645e-07,5.780736e-07,8.428058e-07))
print(tran)
}
