% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GT_tran_retest.R
\name{print.GT_tran_retest}
\alias{print.GT_tran_retest}
\title{Transfer learning in group testing with retesting information}
\usage{
\method{print}{GT_tran_retest}(x, ...)
}
\arguments{
\item{x}{An object of class `GT_tran_retest`.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print a summary of GT_tran_retest.
}
\examples{
tran_retest = GT_tran_retest(500,5,0.95,0.95,0.999,0.999,41,c(23,18,0,0,0,0),
c(0.009627190,0.009347203,0.008905579,0.009290185,0.010037513),
c(9.083533e-07,2.552419e-07,6.768645e-07,5.780736e-07,8.428058e-07))
print(tran_retest)
}
